creating utils file for bondManager so that bondManager can hold the single DotSubstrateCall class.
